feat!(io): Reorganized io module to package module and domain nested module for internals
