+++ threemf2-benches/benches/threemf_reader_multi.rs
@@ -2,7 +2,7 @@ use criterion::{BenchmarkId, Criterion, criterion_group, criterion_main};
 use lib3mf_core::archive::{ArchiveReader, ZipArchiver};
 use serde::Deserialize;
 use std::path::PathBuf;
-use threemf2::io::ThreemfPackage;
+use threemf2::package::ThreemfPackage;
 
 #[cfg(feature = "enable-alloc-check")]
 #[global_allocator]
