+++ threemf2-xsd-validation/tests/production_validation.rs
@@ -5,7 +5,6 @@
 
 use std::io::Cursor;
 use threemf2::{
-    io::ThreemfPackageBuilder,
     model::{
         OptionalResourceId, OptionalResourceIndex, UuidResource,
         domain::{
@@ -17,6 +16,7 @@ use threemf2::{
             resources::Resources,
         },
     },
+    package::ThreemfPackageBuilder,
     threemf_namespaces::ThreemfNamespace,
 };
