+++ threemf2/src/package/mod.rs
@@ -1,35 +1,12 @@
-pub mod content_types;
-pub mod relationship;
+pub mod domain;
 
 mod error;
 pub use error::Error;
-
-mod utils;
-
-pub mod thumbnail_handle;
 
 #[cfg(feature = "io-write")]
 pub mod builder;
 #[cfg(feature = "io-write")]
 pub use builder::ThreemfPackageBuilder;
-
-pub mod validator;
-mod validator_rules;
-
-/// Represents a generic XML namespace declaration with its prefix and URI
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct XmlNamespace {
-    /// The namespace prefix (None for default namespace)
-    pub prefix: Option<String>,
-    /// The namespace URI
-    pub uri: String,
-}
-
-#[cfg(any(
-    feature = "io-memory-optimized-read",
-    feature = "io-speed-optimized-read"
-))]
-mod zip_utils;
 
 #[cfg(any(
     feature = "io-write",
