+++ threemf2/tests/boolean_io.rs
@@ -12,9 +12,9 @@ mod tests {
     #[cfg(feature = "io-memory-optimized-read")]
     #[test]
     fn read_threemf_package_memory_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_boolean_shape_objects;
-        use threemf2::io::query::get_mesh_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_boolean_shape_objects;
+        use threemf2::package::query::get_mesh_objects;
 
         let path = PathBuf::from("./tests/data/mesh-booleans-operations-material.3mf");
         let reader = File::open(path).unwrap();
@@ -103,9 +103,9 @@ mod tests {
     #[cfg(feature = "io-speed-optimized-read")]
     #[test]
     fn read_threemf_package_speed_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_boolean_shape_objects;
-        use threemf2::io::query::get_mesh_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_boolean_shape_objects;
+        use threemf2::package::query::get_mesh_objects;
 
         let path = PathBuf::from("./tests/data/mesh-booleans-operations-material.3mf");
         let reader = File::open(path).unwrap();
@@ -181,7 +181,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-memory-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_memory_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-booleans-operations-material.3mf");
         let reader = File::open(path).unwrap();
@@ -260,7 +260,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-speed-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_speed_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-booleans-operations-material.3mf");
         let reader = File::open(path).unwrap();
