+++ threemf2/tests/core_io.rs
@@ -12,10 +12,10 @@ mod tests {
     #[cfg(feature = "io-memory-optimized-read")]
     #[test]
     fn read_threemf_package_memory_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_components_objects;
-        use threemf2::io::query::get_mesh_objects;
-        use threemf2::io::query::get_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_components_objects;
+        use threemf2::package::query::get_mesh_objects;
+        use threemf2::package::query::get_objects;
 
         let path = PathBuf::from("./tests/data/mesh-composedpart.3mf");
         let reader = File::open(path).unwrap();
@@ -26,7 +26,7 @@ mod tests {
 
         match result {
             Ok(package) => {
-                use threemf2::io::validator::{ValidationRule, Validator};
+                use threemf2::package::domain::validator::{ValidationRule, Validator};
 
                 assert_eq!(package.relationships.len(), 1);
 
@@ -70,10 +70,10 @@ mod tests {
     #[cfg(feature = "io-speed-optimized-read")]
     #[test]
     fn read_threemf_package_speed_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_components_objects;
-        use threemf2::io::query::get_mesh_objects;
-        use threemf2::io::query::get_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_components_objects;
+        use threemf2::package::query::get_mesh_objects;
+        use threemf2::package::query::get_objects;
 
         let path = PathBuf::from("./tests/data/mesh-composedpart.3mf");
         let reader = File::open(path).unwrap();
@@ -84,7 +84,7 @@ mod tests {
 
         match result {
             Ok(package) => {
-                use threemf2::io::validator::{ValidationRule, Validator};
+                use threemf2::package::domain::validator::{ValidationRule, Validator};
 
                 assert_eq!(package.relationships.len(), 1);
 
@@ -128,7 +128,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-memory-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_memory_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-composedpart.3mf");
         let reader = File::open(path).unwrap();
@@ -195,7 +195,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-speed-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_speed_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-composedpart.3mf");
         let reader = File::open(path).unwrap();
