+++ threemf2/tests/material_io.rs
@@ -8,14 +8,14 @@ mod tests {
 
     use std::{fs::File, path::PathBuf};
 
-    use threemf2::io::query::get_mesh_objects;
     use threemf2::model::Color;
     use threemf2::model::query::get_color_groups_from_model;
+    use threemf2::package::query::get_mesh_objects;
 
     #[cfg(feature = "io-memory-optimized-read")]
     #[test]
     fn read_threemf_package_memory_optimized() {
-        use threemf2::io::ThreemfPackage;
+        use threemf2::package::ThreemfPackage;
 
         let path = PathBuf::from("./tests/data/mesh_vertexcolor-material.3mf");
         let reader = File::open(path).unwrap();
@@ -64,7 +64,7 @@ mod tests {
     #[cfg(feature = "io-speed-optimized-read")]
     #[test]
     fn read_threemf_package_speed_optimized() {
-        use threemf2::io::ThreemfPackage;
+        use threemf2::package::ThreemfPackage;
 
         let path = PathBuf::from("./tests/data/mesh_vertexcolor-material.3mf");
         let reader = File::open(path).unwrap();
