+++ threemf2/tests/production_io.rs
@@ -12,10 +12,10 @@ mod tests {
     #[cfg(feature = "io-memory-optimized-read")]
     #[test]
     fn read_threemf_package_memory_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_components_objects;
-        use threemf2::io::query::get_mesh_objects;
-        use threemf2::io::query::get_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_components_objects;
+        use threemf2::package::query::get_mesh_objects;
+        use threemf2::package::query::get_objects;
 
         let path = PathBuf::from("./tests/data/mesh-composedpart-separate-model-files.3mf");
         let reader = File::open(path).unwrap();
@@ -65,11 +65,10 @@ mod tests {
     #[cfg(feature = "io-speed-optimized-read")]
     #[test]
     fn read_threemf_package_speed_optimized() {
-        use threemf2::io::ThreemfPackage;
-        use threemf2::io::query::get_components_objects;
-        use threemf2::io::query::get_mesh_objects;
-        // use threemf2::io::query::get_object_ref_from_id;
-        use threemf2::io::query::get_objects;
+        use threemf2::package::ThreemfPackage;
+        use threemf2::package::query::get_components_objects;
+        use threemf2::package::query::get_mesh_objects;
+        use threemf2::package::query::get_objects;
 
         let path = PathBuf::from("./tests/data/mesh-composedpart-separate-model-files.3mf");
         let reader = File::open(path).unwrap();
@@ -119,7 +118,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-memory-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_memory_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-composedpart-separate-model-files.3mf");
         let reader = File::open(path).unwrap();
@@ -212,7 +211,7 @@ mod tests {
     #[cfg(all(feature = "io-lazy-read", feature = "io-speed-optimized-read"))]
     #[test]
     fn read_threemf_package_lazy_speed_optimized() {
-        use threemf2::io::{CachePolicy, ThreemfPackageLazyReader};
+        use threemf2::package::{CachePolicy, ThreemfPackageLazyReader};
 
         let path = PathBuf::from("./tests/data/mesh-composedpart-separate-model-files.3mf");
         let reader = File::open(path).unwrap();
